Move testkit to util
